Redesign and refactor sonar messages; simplify obstacle detection
Sonar messages (coming from left, center, and right ultrasonic sensors) are no longer transmitted on three separate topics, but instead are stored in a 1x3 ROS multiarray. This facilitates the receipt of all three values simultaneously, eliminating the need for synchronization between values. Obstacles are now classified as being on 1) the right side of the rover, 2) the front/left side of the rover, or 3) not in collision.