Revert redesign of sonar messages due to Gazebo compatibility issue
Combining sonar messages into a single array is not compatible with Gazebo's sonar plugins, which require a unique topic for each sensor. Instead, sonar messages are again transmitted on three separate topics, but we now use ROS TimeSynchronizer to synchronize these topics through a single callback. This synchronizer is a part of the ROS message_filters package, which has now been included in the obstacle package.