Minor redesign of translate state's guard to avoid stuck state
The TRANSLATE state was previously guarded by a check to ensure a minimal distance between the robot's current location and its goal to ensure a correct translation. Because of sensor noise and drift, this check would occasionally allow the robot to become stuck in this state and translate forever. This guard has now been changed to check the angle between the robot's current heading, and the heading from the robot's current position to the goal position. The guard returns FALSE when this angle is at least PI/2, meaning the robot has reached/passed the goal.