Correct target detection and collection
Target collection is now properly implemented in TRANSFORM state. This ensures that a robot that has previously detected a target must return to the center of the arena and clear the targetDetected variable before initiating a new search. Target handler also now checks to see if robot has already detected a tag before recognizing a new one and return home with it.