Track targets across multiple robots as they are collected
Targets are now tracked over the length of a run, such that each target can only be collected once. Robots broadcast their collected targets on the targetsCollected topic, allowing all robots to track these collected targets with the targetsCollected boolean array. For a given target with target ID number n, the target is considered to have already been collected if targetsCollected[n] == 1, and is considered not to have been collected if targetsCollected[n] == 0. The targetDetected variable has also been refactored from an unsigned int to an Int16, and is now initialized to -1 instead of 0xFFFFFFFF.