Fixed a bug where not all the rovers would go into autonomous mode.
The bug was caused by the /rovername/mode topic dissappearing when the selected rover name changed. The all autonomous function changes the selected rover rapidy so the message wasn't reaching all the rovers. I changed the rover mode publisher from a single publisher objec to a map which holds mode publishers for all rovers that have been selected in the past instead of only having one publisher for the currently selected rover. This allows the mode publishers to continue sending data even after the currently selected rover name changes.

I also reenabled the function that stops the joy node when it is not longer being used. This had been disabled for debugging.