Added version and build time display to GUI.
I run two shell commands using CMakeLists.txt at compile time to get the current git version tag and the time at compile. I bind these values to compile time definitions and pass them to the rtq_rover_plugin. These values are displayed in the GUI. The idea is to have an easy way to check that the user is using the code version they expect.