Bug fix: The IMU display rotation matches the simulated rovers
The rover IMU display used to rotate the "wrong" way when compared to the gazebo simulation. I added IMUFrame::inverseRotateByQuaternion() that reverses the direction of rotation. This is accomplished by negating the vector component of the rotation quaternion. This does not explain why the rotation was rotated the wrong way. Possibly because of the view angle in the camera transform, an inversion in the placement of the simulated IMU, or an error in the quaternion rotation code in the GUI.