Bug fix: Intermitent crach when rover selected and sonar display error.
The map display was crashing if ekf odometry data had not been received.
Changed the map to use encoder data for initializing the map coordinate system and check to make sure that at least one position has been received before drawing the map.
Changed the sonar senors in the gazebo model to use 3m as the max range but does not seem to change the values in the topic.
Hardcoded the max and min range to be 3m and 0.01m in the gui display, need to verify these values in the simulator and in the rover.
Fixed a sonar bug where all data was being treated as if it came from the center sonar.