Can now restart the simulation from within the GUI.
This code also helps with handling hardware rovers that connect and disconnect. This is untested though.
Restarting simulations requires that all nodes, subscribers, publishers, and the processes in which they run be properly cleaned up when a simulation is ended.
This cleanup is perfomed in the rqt_rover_plugin and the renamed GazeboSimManager (used to be GazeboSimCreator).
Shutting down all the ros nodes without shutting down the roscore (on which the GUI depends as well) requires that the nodes repspond to SIGINT signals generated by rosnode kill.
The motility node needed a SIGINT handler so it would shutdown when requested.
The EKF and NAVSAT nodes were set to respawn automatically when shutdown in the rover launch files. This was changed to false.

A lot of new code was added and though I have tested it, it will need much more testing since it deals with managing many external processes and the interactions can be unpredictable.
The gazebo client can now be launched and closed from within the GUI independently of the gazebo server. The Rover GUI also catches when the user closes the gazebo client through the gazebo GUI and so properly updates it's internal state.

Once potiential bug I have seen is that when the gazebo server has been restarted a few times the creation of objects slows to a crawl. It will complete and run normally but only after a long time. This has not been reproducable.