Added value filter for joystick control.
Filters values too small to move the physical rover. This was done becasue the joystick will occasionally hang on small values when it is not being touched. This resulted in the motors producing a "whine" while not moving.

The values are filtered for display as well so that the gui displays the same values the motors receive.