Modified GUI to track resource collection according to official rules
The official Swarmathon rules (https://github.com/BCLab-UNM/Swarmathon-Docs/blob/master/Competition%20Rules.md) specify that robots who discover a resources must return to the center collection zone in order to receive credit for finding the resource. Accordingly, the rqt GUI has been updated to ensure that resources are only counted as "collected" if two sequential actions occur: i) a robot publishes a valid image of the AprilTag for the discovered resource on the topic "/targetPickUpImage", and ii) the *same robot* subsequently publishes a valid image of the AprilTag associated with the collection zone on the topic "/targetDropOffImage". The collection zone AprilTags will be tiled across the 50 cm radius area -- these AprilTags *will not* be counted as collected resources, but rather, only as tags that conclude a successful resource collection event.

In order to accomplish the above, the AprilTag library has been added to the rqt_rover_gui package, and the targetDetect and u8 conversion functions from the target package have been duplicated here as well. Future versions of this ROS framework will eliminate this redudancy by combining all necessary AprilTag functionality into a single package.