Implement GUI target value publishers to resolve ambiguity
Users reported discrepancies between the tags reported by the target detection node and those reported by the rqt GUI. As a kludgy patch, we're now publishing the values for the tags officially recorded by the GUI as being "picked up" and "dropped off". Note that the "picked up" value will be equal to the tag detected in the received image with the smallest value that has not already been collected, and the "dropped off" value should always be equal to 256.