My God--It's Full of Cubes!
All 256 tag model SDF files have been updated to be represented
as 3D cubes instead of flat tags.

Functionally, the tag scanning and resource gathering algorithms
remain unchanged and should still function as each cube will
be individually numbered with a unique april tag.

Future updates will deprecate that feature and use a single
april tag number. That will coincide with the development of
collection algorithms that utilize the rover gripper.

In the future, the rovers will be able to grip and move these
cube tags to the nest, but for now, only the shape of the tags
has been updated.