Collapse simulation launch files into single template
Since our ROS launch files receive the robot name as an input argument (i.e. a parameter), we do not require an individual launch file for each simulated rover. We now use a single template launch file in place of multiple, individual files. This update required minimal modification to the rqt GUI to ensure that the simulation ran as expected.