Remove mobility<->GUI target image publish/subscribe framework
As we transition from simulated resource collection to physical pick up and drop off with the gripper, we no longer require the tag scoring infrastructure that was previously implemented to authenticate resource collection through the rqt GUI. The target pick up and drop off publishers and subscribers have been removed from the mobility package, as well as the rqt GUI, meaning that the GUI will no longer maintain a count of collected tags. Additional infrastructure, including the custom TagsImage message and the GUI's AprilTag detection function, will be removed in future commits as we continue this transition.