Replace custom target detection node with preexisting apriltags_ros node
We swap out our custom-built AprilTag processing node for an off-the-shelf ROS package that provides additional functionality, including a pose estimation feature that computes the position of each AprilTag in an image, relative to the position of the camera. AprilTag poses are published in an AprilTagDetectionArray structure, which is received by the mobility node and processed for subsequent action. Note that this array structure will be empty in the case that no tags are detected in the image, therefore we add a guard to mobility's target callback handler to check for this condition.