Implemented single cubes for the Cluster Distribution
The clustered distribution now places 4 x (8 x 8) clusters of
single cubes. The simulation is severely slowed by both the
process of placing the cubes and the general speed of the
simulation itself.

Future work may want to consider optimising the physics engine
to accomodate the many collidable objects now in the model.