Updated cube tags to be static objects
Having the cubes as non-static objects is more accurate for the
simulation, but the strain on the physics engine was too much to run
at a reasonable framerate.

Therefore, the cubes have been changed into static objects so that
the physics engine doesn't have to spend extra processing time
dealing with 256 cubes.

The rovers will still collide with the cubes and the rovers can
interact with the grippers.