Initial implementation of automated cube gripping
The first implementation of autonomous pick up and drop off of AprilTag cubes required several major modifications to the mobility package, including creating a transform listener, replacing old tag detection/collection infrastructure, and adding a tag detection timeout/give-up mechanism.

As before, robots search for tags using a random walk until one (or more) tags are detected. Upon detection, a robot will immediately lower and open its gripper to prepare for pick up, identify the pose of the tag in the odom coordinate frame (offset by the distance from the center of the robot to the center of the gripper fingers), then begin moving towards the target. Positive target detection also triggers a timer that forces the robot to give up moving towards a tag (by switching back to random search) if no tags have been seen for five seconds; the timer is restarted every time a tag is detected. Because robots cannot actually detect tags that are close enough to be picked up, the timer doubles as a pick up mechanism; robots then use the distance from the camera to the tag to identify when a tag has officially been "collected". Once a tag has been collected, the robot then returns to the collection zone and drops off the tag.

Further refinements to this automated behavior are to be expected as testing continues.