Add finger and wrist angle publishers to abridge
The abridge node now publishes the gripper finger and wrist angles that are returned by the Arduino microcontroller. The angle values are precisely the output of the Arduino's read() servo function (converted into radians), which simply returns the last value passed to its write() servo function. Note that unlike the IMU, odometry, and ultrasonic data, these angle values are not truly "sensed," but rather, are read directly out of memory. Accordingly, the topics that publish these angles are /robotName/fingerAngle/prev_cmd and /robotName/wristAngle/prev_cmd, where "prev_cmd" signifies the previous angle command issued to the given servo. The finger and wrist angles are both published as QuaternionStamped for convenience, i.e., for easy access to a built-in, time-stamped orientation type.