Bug fix: change to approximate time synchronizer in obstacle node
Adding timestamps to the sonar data that is being published by abridge, which fits with ROS's recommended style, caused an issue with obstacle detection. Essentially, the obstacle_detection node stopped triggering the sonar callback handler because the timestamps on the left, center, and right sonars never matched *exactly*. As a fix, the time synchronizer is now using an ApproximateTime sync policy, which allows for small variations between timestamps while still triggering callbacks.