updated all six rover SDF model files
to include updated parameters for the camera sensor:

The height value has been changed from 320px to 480px to match
the aspect ratio of the camera for the physical rovers.

The pose of the sensor has been changed to correct an issue
where the simulated rovers had a different field of view from
the physical rovers.