fixed gripper bug in GripperPlugin.cpp
by adding a maximum angle limit for gripping an object. This removes the bug where
a rover would erroneously grip a cube by moving towards it with the gripper fingers
open at its maximum angle. This would cause the cube to become stuck to the gripper.

The gripper fingers can currently open to a maximum angle of 90 degrees. The new
maximum angle limit is set to 80 degrees. Unless the gripper fingers are at 80
degrees or less, the gripper will not be able to grip an object.