corrected an issue with the rover SDF model files
where the rovers would rock back and forth to the point that it
would cause driving problems in the simulated rovers.

Future hotfix or feature updates will implement a more robust
solution, but the fix for now is to adjust the weights of the
wheels and body of the simulated rovers so that they are
heavier and do not bounce around as much when moving back and
forth repeatedly.

The weights may not be perfectly physically accurate, but the
performance of the rovers in simulation now more directly
reflects how the physical rovers behave.