Updated Simulation Timer to differentiate between sim and physical runs
such that the timer is ONLY to be used for the simulated rovers. If
the simulation timer is set when running physical rovers nothing will
occur. If the simulation timer is set when running simulated rovers
appropriate messages will be printed to the infoLogMessage() and the
timer will be initialized.