Redirect rover process output to log file
Added a log file path "app_base/logs" and store the output from spawned rover process to a file. THe file name is the name of the rover with a .log extension.
Added a parameter to the swarmie.launch file so output is sent to the screen - so it can be captured by the redirect from the terminal screed to the log file.
The previous log is overwritten on process launch.
This was done to allow debugging output from nodes running in processes spawned by the simulation.