implemented unique rover color option for the GUI
Bookkeeping variables and related functions have been added
to the MapFrame class to keep track of rover-specific display
colors to be used in the MapFrame display.

With the new checkbox in the GUI, we can swap between the
standard color display for the rover paths and positions on the
MapFrame and rover-unique colors.

The contrast of colors will be useful for debugging and
distinguishing between rovers whose paths cross.