updated the texture image file for the collection zone
Previously, all of the april tags on the collection zone pointed
in the same direction.

With this new update, the april tags have been oriented in such
a manner that they all point inwards towards the center of the
collection zone.

This orientation of tags will allow for programming that will
enable the rovers to intelligently determine if they are inside
of the collection zone, and thus, be able to navigate properly
and avoid getting stuck inside the collection zone.