Bug Fix: No Diag Log for Physical Rovers
The Gazebo environment variables were not set causing the simulated vs physical check in Diagnostics to fail. This resulted in no diagnostic data being sent to the GUI.

Previously those variables were set in the bashrc script. I am trying to move those environment variables from bashrc into the local scripts used to start the nodes. This removes an assumption that scripts are running in a login shell.