Bug fix: Limited the max motor command values
Fixed a problem where the values sent to the Arduino were not capped. The robot turns violently causing the motors to stop working temporarily. Reduced the cap for manual driving to 120 from 255 and applied the cap to autonomous driving where it wasn't applied before. Tested in two physical rovers and in the sim.