Rovers can tell if they are inside a collection zone
Rovers sometimes get stuck inside a collection zone. This is because they avoid running over collection zone AprilTags when not carrying a target.

In this branch the AprilTags are oriented so they all point away from the center. This means a rover can get the yaw values of the tags and determine whether it is inside or outside the collection zone.

If inside, the rover ignores the tags and can drive out of the zone.

If outside, the rover treats the tags as obstacles and tries not to enter the zone unless carrying a target.