Removed obstacle avoidance from Manual mode controllers
Including obstacle avoidance for the sake of manual waypoints has
undesireable effects when controlling the rover with the joystick. It
can also cause idle robots to begin moving when another robot drives
near them. Seems better to just declare the in manual mode obstacle
avoidance is the responsibility of the person operating the robot.