Rationalised the software license The licenses specified in the several package.xml files were inconsistent and in some cases were still just example placeholders.
I have replaced all the licenses specified in the packages with the MIT license and included the MIT license text file in the root of the Swarmathon code directory.

I listed all contributors that have written code and are therefore listed by github as an author. I also listed people who contributed code before it was under git control. Melanie Moses is the PI for the Swarmathon and director of the Biological Computation Lab that houses the Swarmathon project. There may be some students who contributed ideas that we used in the swarmathon code. Apologies to those not included here.

The MIT license is permissive (the only restriction is that the authors are credited in derivative work) and is the most widely used license in in the open-source community.