Make better use of GPS
The odeometry estimate of the collection zone location is synced with the GPS estimate when they differ by more than 0.5 (previously 5m). This change increases the number of cubes collected by 3 rovers in 30 mins from 1 to 13.

The startup delay has been changed from 15 s to 30 s. This is to give the rovers more time to estimate their initial location using GPS.