Refactored to make it possible to test using Mock timers
This change removes the need for any behavior that relies on timers to
be a template class. Instead a pointer to a timer is pased to the
constructor. This means that the timers must be instantiated
independently; however, I believe that is a reasonable price to pay to
make them "mockable."

Some additional refactoring went in to this commit as well:

* timers do not use a callback to modify the behavior that is using
  them. This seems a little cumbersome, but ultimately is a better
  design as if forces the behavior to be more clearly defined (and
  facillitates testing).
* Pickup was refactored to use only a single timer.
* A mock timer class is provided in `test`.