Update interval comment out of sync with the code
The comment described a 1 min interval but the code was 1s. Removed the comment since the interval is expected to be changed by the user as needed. The main intent of the comment was to describe the syntax of the pause command parameter. However, comments describing standard bash usage detract from explaining how this script works.