Added wheel ODE parameters
See PR #243. This pull request supersedes the earlier one since we are using templated model files now. Thanks to @darrenchurchill. 

"I settled on a very small Constraint Force Mixing parameter (0.01), to soften the wheel joints slightly, and a greatly increased the Error Reduction Parameter (0.8), so the wheels move quickly back to their normal positions in just a few physics iterations if they get slightly out of place.

The CFM is small enough so there shouldn't be any 'suspension' effect or vibration in the model when it comes to a stop."