Add sonar plugin from outstanding PR #697 on ros-simulation/gazebo_ros_pkgs
Thanks to @mbennehar

https://github.com/ros-simulation/gazebo_ros_pkgs/pull/697

This is modeled after the range sensor plugin, but by using the sonar sensor, beams hitting the ground don't get reflected back to the sensor. This makes the sensor much more realistic, allowing its field of view to be larger without causing false ground readings. With the increased field of view, objects on the ground will sometimes reflect the beam.

Co-authored-by: Moussab Bennehar <mbennehar@users.noreply.github.com>